Extract some now shared components between AssignmentProgress.tsx and QuizTeacherFeedback.tsx
