+++ src/app/components/pages/quizzes/QuizTeacherFeedback.tsx
@@ -6,32 +6,26 @@ import * as RS from "reactstrap";
 import {ShowLoading} from "../../handlers/ShowLoading";
 import {
     loadQuizAssignmentFeedback,
-    returnQuizToStudent, updateQuizAssignmentDueDate,
+    updateQuizAssignmentDueDate,
     updateQuizAssignmentFeedbackMode
 } from "../../../state/actions/quizzes";
 import {selectors} from "../../../state/selectors";
 import {TitleAndBreadcrumb} from "../../elements/TitleAndBreadcrumb";
-import {
-    IsaacQuizSectionDTO,
-    Mark,
-    QuizAssignmentDTO,
-    QuizFeedbackMode,
-    QuizUserFeedbackDTO
-} from "../../../../IsaacApiTypes";
-import {AssignmentProgressLegend, formatMark, ICON} from '../AssignmentProgress';
+import {QuizFeedbackMode} from "../../../../IsaacApiTypes";
+import {AssignmentProgressLegend} from '../AssignmentProgress';
 import {usePageSettings} from "../../../services/progress";
-import {PageSettings, QuizFeedbackModes} from "../../../../IsaacAppTypes";
+import {QuizFeedbackModes} from "../../../../IsaacAppTypes";
 import {teacherQuizzesCrumbs} from "../../elements/quiz/QuizAttemptComponent";
 import {extractTeacherName} from "../../../services/user";
 import {isDefined} from "../../../services/miscUtils";
 import {formatDate} from "../../elements/DateString";
 import {Spacer} from "../../elements/Spacer";
-import {isQuestion} from "../../../services/questions";
 import {API_PATH} from "../../../services/constants";
-import {closeActiveModal, getQuizAssignmentResultsSummaryCSV, openActiveModal, showToast} from "../../../state/actions";
+import {getQuizAssignmentResultsSummaryCSV, showToast} from "../../../state/actions";
 import {IsaacSpinner} from "../../handlers/IsaacSpinner";
 import {currentYear, DateInput} from "../../elements/inputs/DateInput";
 import {range} from "lodash";
+import { ResultsTable } from "../../elements/quiz/QuizProgressCommon";
 
 interface QuizTeacherFeedbackProps {
     match: {params: {quizAssignmentId: string}}
@@ -40,165 +34,6 @@ interface QuizTeacherFeedbackProps {
 const pageHelp = <span>
     See the feedback for your students for this test assignment.
 </span>;
-
-interface ResultsTableProps {
-    assignment: QuizAssignmentDTO;
-    pageSettings: PageSettings;
-}
-
-function questionsInSection(section?: IsaacQuizSectionDTO) {
-    return section?.children?.filter(child => isQuestion(child)) || [];
-}
-
-const passMark = 0.75;
-function markQuestionClasses(row: QuizUserFeedbackDTO, mark: Mark | undefined, totalOrUndefined: number | undefined) {
-    if (!row.user?.authorisedFullAccess) {
-        return "revoked";
-    }
-
-    const correct = mark?.correct as number;
-    const incorrect = mark?.incorrect as number;
-    const total = totalOrUndefined as number;
-
-    if (correct === total) {
-        return "completed";
-    } else if ((correct / total) >= passMark) {
-        return "passed";
-    } else if ((incorrect / total) > (1 - passMark)) {
-        return "failed";
-    } else if (correct > 0 || incorrect > 0) {
-        return "in-progress";
-    } else {
-        return "not-attempted";
-    }
-}
-
-interface ResultRowProps {
-    pageSettings: PageSettings;
-    row: QuizUserFeedbackDTO;
-    assignment: QuizAssignmentDTO;
-}
-
-function ResultRow({pageSettings, row, assignment}: ResultRowProps) {
-    const [dropdownOpen, setDropdownOpen] = useState(false);
-    const [working, setWorking] = useState(false);
-    const dispatch = useDispatch();
-
-    const toggle = () => setDropdownOpen(prevState => !prevState);
-
-    const returnToStudent = async () => {
-        dispatch(openActiveModal({
-            closeAction: () => {
-                dispatch(closeActiveModal())
-            },
-            title: "Allow another attempt?",
-            body: "This will allow the student to attempt the test again.",
-            buttons: [
-                <RS.Button key={1} color="primary" outline target="_blank" onClick={() => {dispatch(closeActiveModal())}}>
-                    Cancel
-                </RS.Button>,
-                <RS.Button key={0} color="primary" target="_blank" onClick={_returnToStudent}>
-                    Confirm
-                </RS.Button>,
-        ]
-        }));    
-    }
-
-    const _returnToStudent = async () => {
-        try {
-            setWorking(true);
-            await dispatch(returnQuizToStudent(assignment.id as number, row.user?.id as number));
-        } finally {
-            setWorking(false);
-            dispatch(closeActiveModal());
-        }
-    };
-
-    const quiz = assignment?.quiz;
-    const sections: IsaacQuizSectionDTO[] = quiz?.children || [];
-
-    let message;
-    if (!row.user?.authorisedFullAccess) {
-        message = "Not sharing";
-    } else if (!row.feedback?.complete) {
-        message = "Not completed";
-    }
-    const valid = message === undefined;
-    return <tr className={`${row.user?.authorisedFullAccess ? "" : " not-authorised"}`} title={`${row.user?.givenName + " " + row.user?.familyName}`}>
-        <th className="student-name">
-            {valid ?
-                <>
-                    <RS.Button color="link" onClick={toggle} disabled={working}>
-                        <div
-                            tabIndex={0}
-                            className="btn quiz-student-menu"
-                            data-toggle="dropdown"
-                            aria-expanded={dropdownOpen}
-                        >
-                            {row.user?.givenName}
-                            <span className="d-none d-lg-inline"> {row.user?.familyName}</span>
-                            <span className="quiz-student-menu-icon">
-                            {working ? <IsaacSpinner size="sm" /> : <img src="/assets/menu.svg" alt="Menu" />}
-                        </span>
-                        </div>
-                    </RS.Button>
-                    {!working && dropdownOpen && <div className="py-2 px-3">
-                        <RS.Button size="sm" onClick={returnToStudent}>Allow another attempt</RS.Button>
-                    </div>}
-                </>
-            :   <>
-                    {row.user?.givenName}
-                    <span className="d-none d-lg-inline"> {row.user?.familyName}</span>
-                </>
-            }
-        </th>
-        {!valid && <td colSpan={sections.map(questionsInSection).flat().length + 1}>{message}</td>}
-        {valid && <>
-            {sections.map(section => {
-                const mark = row.feedback?.sectionMarks?.[section.id as string];
-                const outOf = quiz?.sectionTotals?.[section.id as string];
-                return questionsInSection(section).map(question => {
-                    const questionMark = row.feedback?.questionMarks?.[question.id as string] || {} as Mark;
-                    const icon =
-                        questionMark.correct === 1 ? ICON.correct :
-                        questionMark.incorrect === 1 ? ICON.incorrect :
-                        /* default */ ICON.notAttempted;
-                    return <td key={question.id} className={markQuestionClasses(row, mark, outOf)}>
-                        {icon}
-                    </td>
-                }).flat()
-            })}
-            <td className="total-column">
-                {formatMark(row.feedback?.overallMark?.correct as number, quiz?.total as number, pageSettings.formatAsPercentage)}
-            </td>
-        </>}
-    </tr>;
-}
-
-function ResultsTable({assignment, pageSettings}: ResultsTableProps) {
-    const sections: IsaacQuizSectionDTO[] = assignment.quiz?.children || [];
-
-    return <table className="progress-table w-100 mb-5 border">
-        <tbody>
-            <tr className="bg-white">
-                <th>&nbsp;</th>
-                {sections.map(section => <th key={section.id} colSpan={questionsInSection(section).length} className="border font-weight-bold">
-                    {section.title}
-                </th>)}
-                <th rowSpan={2} className="border-bottom">Overall</th>
-            </tr>
-            <tr className="bg-white">
-                <th className="bg-white border-bottom">&nbsp;</th>
-                {sections.map(section => questionsInSection(section).map((question, index) => <th key={question.id} className="border">
-                    {`Q${index + 1}`}
-                </th>)).flat()}
-            </tr>
-            {assignment.userFeedback?.map(row =>
-                <ResultRow key={row.user?.id} pageSettings={pageSettings} row={row} assignment={assignment} />
-            )}
-        </tbody>
-    </table>;
-}
 
 const feedbackNames: Record<QuizFeedbackMode, string> = {
     NONE: "No feedback for students",
